API

ProbabilityBoundsAnalysis.pbox — Method

pbox( x :: Array{Interval{T}, 1} ) where T <: Real

Constructs a pbox from an array of intervals with equal mass. Left and right bounds are sorted to construct cdf bounds.

ProbabilityBoundsAnalysis.pbox — Method

pbox( x :: Array{T, 2} ) where T <: Real

Constructs a pbox from a matrix of 2nd order samples [Nouter x Ninner]

Distributions.cdf — Method

cdf(s :: pbox, x :: Real)

Returns cdf value (interval) at x

See also: cut, mass, rand

Distributions.cdf — Method

cdf(s :: pbox, x :: Interval)

Returns interval bounds on cdf value in interval x

See also: cut, mass, rand

ProbabilityBoundsAnalysis.mass — Method

mass(s :: pbox, lo :: Real, hi :: Real)

Returns bounds on probability mass in interval [lo, hi]

See also: cut, cdf, rand

ProbabilityBoundsAnalysis.mass — Method

mass(s :: pbox, x:: Interval)

Returns bounds on probability mass in interval x

See also: cut, cdf, rand

ProbabilityBoundsAnalysis.makepbox — Method

makepbox(x...)

Returns an array of pboxes from an array of inputs (eg an array of intervals or reals).

Examples

julia> s = makepbox(interval(0,1))
 Pbox: 	  ~  ( range=[0.0,1.0], mean=[0.0,1.0], var=[0.0,0.25])

julia> ar = [interval(0, 1), interval(0, 2), 3];
julia> s = makepbox.(ar)
3-element Array{pbox,1}:
 Pbox: 	  ~  ( range=[0.0,1.0], mean=[0.0,1.0], var=[0.0,0.25])
 Pbox: 	  ~  ( range=[0.0,2.0], mean=[0.0,2.0], var=[0.0,1.0])
 Pbox: 	  ~  ( range=[-1.0,0.0], mean=[-1.0,0.0], var=[0.0,0.25])

Statistics.mean — Method

mean( x :: pbox)

Get interval mean of a pbox See also: var, std

Statistics.var — Method

var( x :: pbox)

Get interval variance of a pbox See also: std, mean

Statistics.std — Method

std( x :: pbox)

Get interval std of a pbox See also: var, mean

ProbabilityBoundsAnalysis.env — Method

env(x :: pbox, y :: pbox, ...)

Envelope. Returns the union of pboxes. Any number of pboxes may be input

Examples

julia> a = U(0,1)

julia> b = U(1,2)

julia> c = env(a, b)
Pbox: 	  ~ uniform ( range=[0.0,2.0], mean=[0.5,1.5], var=0.08333333333333333)

See also: imp, makepbox

ProbabilityBoundsAnalysis.imp — Method

imp(x :: pbox, y :: pbox)

Imprint. Returns the intersection of pboxes. Any number of pboxes may be input

Examples

julia> a = U(interval(0, 1), 2)

julia> b = U(1, 2)

julia> c = imp(a, b)
Pbox: 	  ~  ( range=[1.0,2.0], mean=1.5, var=0.08333333333333333)

See also: env, makepbox

ProbabilityBoundsAnalysis.normal — Method

normal(mean :: Interval, std :: Interval)

Normal shaped pbox. Parameters can be Real or Intervals.

Constructors

normal

N

gaussian

Examples

julia> a = normal(interval(0, 1), interval(1,2))
Pbox: 	  ~ normal ( range=[-6.1805, 7.1805], mean=[0.0, 1.0], var=[1.0, 4.0])

See also: uniform, lognormal, meanMinMax, plot

Missing docstring.

Missing docstring for uniform(min :: Interval, max :: Interval). Check Documenter's build log for details.

ProbabilityBoundsAnalysis.beta — Method

beta(α :: Interval, β :: Interval)

Beta shaped pbox. Parameters can be Real or Intervals.

Examples

julia> a = beta(2,interval(3,4))
Pbox: 	  ~ beta ( range=[0.0, 1.0], mean=[0.33333, 0.4], var=[0.031746, 0.04])

See also: KN, meanMinMax, plot

ProbabilityBoundsAnalysis.lognormal — Method

lognormal(μ :: Interval, std :: Interval)

Lognormal shaped pbox. Parameters can be Real or Intervals.

See also: KN, meanMinMax

ProbabilityBoundsAnalysis.KN — Method

KN(k :: Interval, n :: Interval)

k out of N confidence box (c-box), a pbox shaped confidence structure. Quantifies inferential uncertainty in binomial counts, where k successes were observed out of n trails. One sided or two sided confidence intervals may be drawn

Constructors

KN

kn

See also: meanMinMax, plot

ProbabilityBoundsAnalysis.cut — Method

cut(x :: pbox, p :: Real)

returns a vertical cut of a pbox at cdf value p, for p ∈ [0, 1]

Constructors

cut(x :: pbox, p :: Real)

cut(x :: pbox, p :: Interval)

See also: rand, cdf, mass

Index

ProbabilityBoundsAnalysis.pbox

ProbabilityBoundsAnalysis.pbox

Distributions.cdf

Distributions.cdf

ProbabilityBoundsAnalysis.KN

ProbabilityBoundsAnalysis.beta

ProbabilityBoundsAnalysis.cut

ProbabilityBoundsAnalysis.env

ProbabilityBoundsAnalysis.imp

ProbabilityBoundsAnalysis.lognormal

ProbabilityBoundsAnalysis.makepbox

ProbabilityBoundsAnalysis.mass

ProbabilityBoundsAnalysis.mass

ProbabilityBoundsAnalysis.normal

Statistics.mean

Statistics.std

Statistics.var